+++ crates/relayer/src/link.rs
@@ -33,8 +33,8 @@ pub use relay_path::{RelayPath, Resubmit};
 pub struct LinkParameters {
     pub src_port_id: PortId,
     pub src_channel_id: ChannelId,
-    pub max_memo_size: usize,
-    pub max_receiver_size: usize,
+    pub max_memo_size: u64,
+    pub max_receiver_size: u64,
 }
 
 pub struct Link<ChainA: ChainHandle, ChainB: ChainHandle> {
@@ -45,8 +45,8 @@ impl<ChainA: ChainHandle, ChainB: ChainHandle> Link<ChainA, ChainB> {
     pub fn new(
         channel: Channel<ChainA, ChainB>,
         with_tx_confirmation: bool,
-        max_memo_size: usize,
-        max_receiver_size: usize,
+        max_memo_size: u64,
+        max_receiver_size: u64,
     ) -> Result<Self, LinkError> {
         Ok(Self {
             a_to_b: RelayPath::new(
